Adding third validation js
Feat: adding the third validation

The attribs object passed to the validate function represents what the
attributes would be after completing the current set() or save(). This
object is distance from the current attributes of the model and form
the parameters passed to the operation. Since it is created by shallow
copy, it is not possible to change any Number, String, or Boolean
attribute of the input with function, but it is possible to change
attributes in nested objects