using set() to update the collection
Feat: Using set() method to update collection

The set() method available for Collections can also be used for “smart”
updating of sets of models. This method attempts to perform smart
updating of a collection using a specified list of models. When a model
in this list isn’t present in the collection, it is added. If it’s
present, its attributes will be merged. Models which are present in the
collection but not in the list are removed.