Fetching & Saving Models to and from Server
Feat: fetch and save models to and from server

In order to fetch the models from server, use Collection.fetch() to
thieves a set of models from the server in a form os a JSON array by
sending an HTTP GET request to the URL specified by the collection’s
url property. When the data is received, a set() will be executed to
update the collection.

Use save() method on the model to updates models. When save() is called
on a model that was fetched from the server, it constructs a URL by
appending the mode’s id to the collection’s URL and sends an HTTP PUT
to the server. If the model is a new instance that was created in the
browser(i.e it doesn’t have an id ) then an HTTP POST is sent to the
collection’URL.